backend/requirements.txt
fastapi
uvicorn
sqlalchemy
aiomysql
pydantic
pydantic-settings
python-multipart
pytest
pytest-asyncio
httpx
cryptography
python-jose[cryptography]
passlib[bcrypt]
greenlet
numpy
pytz
redis